chore: rename singleCreate route to create
